requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
